requirements.txt
aiogram==3.20.0.post0
python-dotenv==1.1.0
aiohttp==3.11.18
asyncpg==0.30.0
sqlalchemy==2.0.40
apscheduler==3.11.0